refactor(demo): collocate created components
