+++ apps/demo/src/cms.tsx
@@ -2,6 +2,7 @@ import { contentPaths, initContent } from '@/content'
 import { createJsonResolver } from '@deepdish/resolvers/json'
 import { createComponents } from '@deepdish/ui/components'
 import { configure } from '@deepdish/ui/config'
+import type { DeepDishProps } from '@deepdish/ui/deepdish'
 import * as v from 'valibot'
 
 const featureSchema = v.object({
@@ -45,5 +46,28 @@ async function cms() {
 
 const components = await cms()
 
-export const Feature = components.feature
 export const Text = components.text
+
+export function Feature(props: {
+  deepdish: DeepDishProps
+}) {
+  const Component = components.feature
+
+  return (
+    <Component
+      deepdish={props.deepdish}
+      render={async (value) => {
+        return (
+          <div className="border border-gray-200 rounded-lg p-4 bg-white shadow-sm inline-block">
+            <h1 className="text-xl text-gray-800 font-bold">
+              {value?.name ?? 'Feature Name'}
+            </h1>
+            <p className="text-gray-500">
+              {value?.description ?? 'Feature Description'}
+            </p>
+          </div>
+        )
+      }}
+    />
+  )
+}
